mostly broken ProfilePageHeader
